+++ src/client/srcdiff_input_source_local.cpp
@@ -428,7 +428,7 @@ srcdiff_input_source_local::input_context * srcdiff_input_source_local::open(con
 
 }
 
-int srcdiff_input_source_local::read(void * context, char * buffer, int len) {
+int srcdiff_input_source_local::read(void * context, char * buffer, size_t len) {
 
   input_context * ctx = (input_context *)context;
