+++ src/client/srcdiff_input_source_svn.cpp
@@ -571,7 +571,7 @@ srcdiff_input_source_svn::input_context * srcdiff_input_source_svn::open(const c
 
 }
 
-int srcdiff_input_source_svn::read(void * context, char * buffer, int len) {
+int srcdiff_input_source_svn::read(void * context, char * buffer, size_t len) {
 
   input_context * ctx = (input_context *)context;
